pyproject.toml
[project]
name = "tool-calling-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
]
